--- src/pages/Home.js
@@ -1,8 +1,0 @@
-
-export default function Home() {
-    return (
-        <div>
-            
-        </div>
-    )
-}
