+++ src/pages/Login.js
@@ -1,0 +1,8 @@
+
+export default function Login() {
+    return (
+        <div>
+            Hellow From Login
+        </div>
+    )
+}
